Refactor pagination and delete todo logic

+++ lib/features/todo/data/repositories/todo_repository_impl.dart
@@ -5,7 +5,6 @@ import 'package:todo/core/service/api_holder.dart';
 import 'package:todo/core/service/local/objectbox_service.dart';
 import 'package:todo/core/service/network_connectivity.dart';
 import 'package:todo/features/todo/domain/entities/auth.dart';
-import 'package:todo/features/todo/domain/entities/todo.dart';
 import 'package:todo/features/todo/domain/repositories/todo_repository.dart';
 
 import '../../../../core/entities/todo_response.dart';
@@ -21,62 +20,37 @@ class TodoRepositoryImpl implements TodoRepository {
     required this.objectboxService,
     required this.networkConnectivity,
   });
-  int offsetTodo = 0;
-  int offsetDoing = 0;
-  int offsetDone = 0;
   int limit = 10;
-  String sortBy = "createdAt";
-  bool isAsc = true;
-  // bool isFinalPage = false;
   bool isFirstTime = true;
 
   @override
-  Future<Result<GetTodoListRes, TodoError>> getTodoList(
-      TodoStatus status) async {
+  Future<Result<TodoResponse, TodoError>> getTodoList(
+      GetTodoListParam param) async {
     try {
+      if (param.fetchLocalOnly) {
+        return Result.success(TodoResponse(
+            tasks: objectboxService.todos
+                .where((todo) => todo.status == param.status)
+                .toList(),
+            pageNumber: 1,
+            totalPages: 1));
+      }
+      //make clean print of all param data
+      print('status: ${param.status}');
+      print('offset: ${param.offset}');
       if (await networkConnectivity.status) {
-        int offset = 0;
-        switch (status) {
-          case TodoStatus.TODO:
-            offset = offsetTodo;
-            break;
-          case TodoStatus.DOING:
-            offset = offsetDoing;
-            break;
-          case TodoStatus.DONE:
-            offset = offsetDone;
-            break;
-          default:
-            offset = offsetTodo;
-        }
         var result =
             await networkExecuter.execute<TodoResponseModel, TodoResponse>(
           route: ApiHolder.fetchTodoList(
-              status: status,
-              offset: offset,
-              limit: limit,
-              sortBy: sortBy,
-              isAsc: isAsc),
+              status: param.status,
+              offset: param.offset,
+              limit: param.limit,
+              sortBy: param.sortBy,
+              isAsc: param.isAsc),
           responseType: TodoResponseModel(),
         );
 
         return result.when(success: (data) {
-          if (data.tasks.isNotEmpty) {
-            switch (status) {
-              case TodoStatus.TODO:
-                offsetTodo++;
-                break;
-              case TodoStatus.DOING:
-                offsetDoing++;
-                break;
-              case TodoStatus.DONE:
-                offsetDone++;
-                break;
-              default:
-                offsetTodo++;
-            }
-          }
-
           //clean all local data when first request for sync data between local and server
           if (isFirstTime) {
             objectboxService.deleteTodos();
@@ -85,19 +59,21 @@ class TodoRepositoryImpl implements TodoRepository {
 
           objectboxService.saveTodos(data.tasks);
           final todos = objectboxService.todos
-              .where((todo) => todo.status == status)
+              .where((todo) => todo.status == param.status)
               .toList();
-          if (isAsc) {
+          if (param.isAsc) {
             todos.sort((a, b) => a.createdAt.compareTo(b.createdAt));
           }
-          return Result.success(GetTodoListRes(
-              isFinalPage: data.pageNumber == data.totalPages, todos: todos));
+          return Result.success(TodoResponse(
+              pageNumber: data.pageNumber,
+              totalPages: data.totalPages,
+              tasks: todos));
         }, failure: (TodoError error) {
           return Result.failure(error);
         });
       } else {
-        return Result.success(
-            GetTodoListRes(isFinalPage: true, todos: objectboxService.todos));
+        return Result.success(TodoResponse(
+            tasks: objectboxService.todos, pageNumber: 1, totalPages: 1));
       }
     } on TodoError catch (e) {
       return Result.failure(e);
